// src/modules/tableau-electrique/models/piece.enum.ts
export enum Piece {
  SALON = 'Salon',
  CUISINE = 'Cuisine',
  ENTREE = 'Entrée',
  SALLE_D_EAU = `Salle d'eau`,
  TERASSE = 'Terasse',
  GARAGE = 'Garage',
  BUREAU = 'Bureau',
  SALLE_DE_BAIN = 'Salle de bain',
  PALIER_ESCALIER = 'Palier / Escalier',
  CHAMBRE_LEONORE = 'Chambre Léonore',
  CHAMBRE_VALENTIN_LAURENE = 'Chambre V&L',
  TOILETTE = 'Toilette',
  SAS = 'Sas'
}